+++ plugin/CHANGELOG.md
@@ -7,8 +7,8 @@ Change Log
 - Support instrumentation with JUnit 5 and 6 (the plugin will choose the correct runtime accordingly)
 - Introduce `de.mannodermaus.android-junit-framework` as the new plugin ID
 - New: Control behavior of test execution on unsupported devices via `instrumentationTests.behaviorForUnsupportedDevices`
-  - "Skip": Skip tests and mark them as ignored (default and equal to previous behavior)
-  - "Fail": Throw an exception and fail test execution
+  - "Fail": Throw an exception and fail test execution (**this is the new default**)
+  - "Skip": Skip tests and mark them as ignored (**this is the old behavior**)
 
 ## 1.14.0.0 (2025-10-10)
 - JUnit 5.14.0